+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Upcoming
+### Added
+- (Drupal) Add default variables as in template_preprocess().  
+
 ## 1.0.6
 ## 1.0.5
 ### Added